Adds e2e tests for filling out the considerations of views section of the vasu document
