+++ frontend/src/e2e-playwright/pages/employee/vasu/vasu.ts
@@ -4,50 +4,7 @@
 
 import { waitUntilFalse } from 'e2e-playwright/utils'
 import { Page, TextInput, Element } from '../../../utils/page'
-
-export class AuthorsSection extends Element {
-  readonly #primaryInputs = this.findAll(
-    '[data-qa="multi-field-question"] input'
-  )
-  primaryFirstNameInput = new TextInput(this.#primaryInputs.nth(0))
-  primaryLastNameInput = new TextInput(this.#primaryInputs.nth(1))
-  primaryTitleInput = new TextInput(this.#primaryInputs.nth(2))
-  primaryPhoneNumberInput = new TextInput(this.#primaryInputs.nth(3))
-
-  readonly #otherFieldsRows = this.findAll(
-    '[data-qa="multi-field-list-question"] [data-qa="field-row"]'
-  )
-  readonly otherFieldsInputs = (ix: number) =>
-    this.#otherFieldsRows.nth(ix).findAll('input')
-  otherFirstNameInput = (ix: number) =>
-    new TextInput(this.otherFieldsInputs(ix).nth(0))
-  otherLastNameInput = (ix: number) =>
-    new TextInput(this.otherFieldsInputs(ix).nth(1))
-  otherTitleInput = (ix: number) =>
-    new TextInput(this.otherFieldsInputs(ix).nth(2))
-  otherPhoneNumberInput = (ix: number) =>
-    new TextInput(this.otherFieldsInputs(ix).nth(3))
-
-  readonly #primaryValue = this.findAll(
-    '[data-qa="multi-field-question"] [data-qa="value-or-no-record"]'
-  ).first()
-
-  readonly #otherFieldsValues = this.findAll(
-    '[data-qa="value-or-no-record"]'
-  ).nth(1)
-
-  get primaryValue(): Promise<string> {
-    return this.#primaryValue.innerText
-  }
-
-  get otherFieldsCount(): Promise<number> {
-    return this.#otherFieldsRows.count()
-  }
-
-  get otherValues(): Promise<string> {
-    return this.#otherFieldsValues.innerText
-  }
-}
+import { AuthorsSection, ConsiderationsSection } from './pageSections'
 
 class VasuPageCommon {
   constructor(readonly page: Page) {}
@@ -59,7 +16,7 @@ class VasuPageCommon {
     '[data-qa="vasu-followup-question"]'
   )
 
-  getDocumentSection(ix: number) {
+  getDocumentSection(ix: number): Element {
     // Note: indexes might change if the template used in the test changes
     return this.#documentSection.nth(ix)
   }
@@ -70,6 +27,10 @@ class VasuPageCommon {
 
   get authorsSection(): AuthorsSection {
     return new AuthorsSection(this.getDocumentSection(0))
+  }
+
+  get considerationsSection(): ConsiderationsSection {
+    return new ConsiderationsSection(this.getDocumentSection(1))
   }
 
   get followupQuestionCount(): Promise<number> {
